Extract UpNext screen to a dedicated package
