+++ ios/Packages/UpNext/Sources/UpNext/UpNextListItemView.swift
@@ -3,62 +3,75 @@ import DesignSystem
 import Models
 import SwiftUI
 
-public struct WatchListItemView: View {
+public struct UpNextListItemView: View {
     @Environment(\.appTheme) private var theme
 
     let episode: SwiftNextEpisode
-    let premiereLabel: String
-    let newLabel: String
     let onItemClicked: (Int64, Int64) -> Void
     let onShowTitleClicked: (Int64) -> Void
     let onMarkWatched: () -> Void
+    let onLongPress: () -> Void
     let isUpdating: Bool
+
+    private let posterImageUrl: String?
+    private let episodeInfoText: String
 
     public init(
         episode: SwiftNextEpisode,
-        premiereLabel: String,
-        newLabel: String,
         onItemClicked: @escaping (Int64, Int64) -> Void,
         onShowTitleClicked: @escaping (Int64) -> Void,
         onMarkWatched: @escaping () -> Void,
+        onLongPress: @escaping () -> Void = {},
         isUpdating: Bool = false
     ) {
         self.episode = episode
-        self.premiereLabel = premiereLabel
-        self.newLabel = newLabel
         self.onItemClicked = onItemClicked
         self.onShowTitleClicked = onShowTitleClicked
         self.onMarkWatched = onMarkWatched
+        self.onLongPress = onLongPress
         self.isUpdating = isUpdating
+
+        posterImageUrl = episode.imageUrl
+
+        var text = episode.episodeNumber
+        if episode.remainingEpisodes > 0 {
+            text += " +\(episode.remainingEpisodes)"
+        }
+        if let runtime = episode.runtime {
+            text += " (\(runtime))"
+        }
+        episodeInfoText = text
     }
 
     public var body: some View {
-        Button(action: {
-            onItemClicked(episode.showTraktId, episode.episodeId)
-        }) {
-            HStack(alignment: .top, spacing: 0) {
-                posterView
-                episodeDetails
-                watchedButton
-            }
-            .frame(height: WatchListItemViewConstants.height)
-            .frame(maxWidth: .infinity)
-            .background(.appSurface)
-            .cornerRadius(WatchListItemViewConstants.cornerRadius)
+        HStack(alignment: .top, spacing: 0) {
+            posterView
+            episodeDetails
+            watchedButton
         }
-        .buttonStyle(PlainButtonStyle())
-        .padding(.horizontal, theme.spacing.medium)
+        .frame(height: UpNextListItemViewConstants.height)
+        .frame(maxWidth: .infinity)
+        .background(.appSurface)
+        .cornerRadius(UpNextListItemViewConstants.cornerRadius)
+        .contentShape(Rectangle())
+        .onTapGesture {
+            onItemClicked(episode.showTraktId, episode.episodeId)
+        }
+        .onLongPressGesture {
+            onLongPress()
+        }
+        .padding(.horizontal, theme.spacing.xSmall)
     }
 
     private var posterView: some View {
         PosterItemView(
             title: nil,
-            posterUrl: episode.imageUrl,
-            posterWidth: WatchListItemViewConstants.imageWidth,
-            posterHeight: WatchListItemViewConstants.height,
+            posterUrl: posterImageUrl,
+            posterWidth: UpNextListItemViewConstants.imageWidth,
+            posterHeight: UpNextListItemViewConstants.height,
             posterRadius: 0
         )
-        .frame(width: WatchListItemViewConstants.imageWidth, height: WatchListItemViewConstants.height)
+        .frame(width: UpNextListItemViewConstants.imageWidth, height: UpNextListItemViewConstants.height)
         .clipped()
     }
 
@@ -69,43 +82,42 @@ public struct WatchListItemView: View {
                 onTap: { onShowTitleClicked(episode.showTraktId) }
             )
 
-            HStack(spacing: theme.spacing.xxSmall) {
-                Text(episode.episodeNumber)
-                    .textStyle(theme.typography.labelLarge)
-                    .foregroundStyle(.appOnSurface)
-                    .lineLimit(1)
-                if episode.remainingEpisodes > 0 {
-                    Text("+\(episode.remainingEpisodes)")
-                        .textStyle(theme.typography.labelMedium)
-                        .foregroundStyle(.appOnSurface.opacity(0.6))
-                }
-            }
-            .padding(.top, theme.spacing.medium)
+            Text(episodeInfoText)
+                .textStyle(theme.typography.bodySmall)
+                .foregroundStyle(.appAccent)
+                .lineLimit(1)
+                .padding(.top, theme.spacing.xSmall)
 
             Text(episode.episodeTitle)
                 .textStyle(theme.typography.bodySmall)
                 .foregroundStyle(.appOnSurface.opacity(0.7))
                 .lineLimit(2)
-                .padding(.top, theme.spacing.xxSmall)
 
             Spacer()
 
-            badgeView
+            progressView
         }
         .frame(maxWidth: .infinity, alignment: .leading)
         .padding(.vertical, theme.spacing.small)
         .padding(.horizontal, theme.spacing.xSmall)
     }
 
-    private var badgeView: some View {
-        HStack(spacing: theme.spacing.xxSmall) {
-            switch episode.badge {
-            case .premiere:
-                PremiereBadge(text: premiereLabel)
-            case .new:
-                NewBadge(text: newLabel)
-            case .none:
-                EmptyView()
+    private var progressView: some View {
+        let progress: Float = episode.totalCount > 0
+            ? Float(episode.watchedCount) / Float(episode.totalCount)
+            : 0
+
+        return HStack(spacing: theme.spacing.xSmall) {
+            SegmentedProgressBar(
+                segmentProgress: [progress],
+                height: 4
+            )
+            .frame(maxWidth: .infinity)
+
+            if episode.totalCount > 0 {
+                Text("\(episode.watchedCount)/\(episode.totalCount)")
+                    .textStyle(theme.typography.labelSmall)
+                    .foregroundStyle(.appOnSurface.opacity(0.6))
             }
         }
     }
@@ -115,35 +127,41 @@ public struct WatchListItemView: View {
             ZStack {
                 Circle()
                     .fill(.appGrey)
-                    .frame(width: WatchListItemViewConstants.checkmarkSize, height: WatchListItemViewConstants.checkmarkSize)
+                    .frame(
+                        width: UpNextListItemViewConstants.checkmarkSize,
+                        height: UpNextListItemViewConstants.checkmarkSize
+                    )
                 if isUpdating {
                     ProgressView()
                         .progressViewStyle(.circular)
                         .tint(theme.colors.onPrimary)
                 } else {
                     Image(systemName: "checkmark")
-                        .textStyle(theme.typography.labelMedium)
-                        .foregroundStyle(.appOnPrimary)
+                        .textStyle(theme.typography.titleSmall)
+                        .foregroundStyle(.white)
                 }
             }
+            .frame(width: UpNextListItemViewConstants.tapTargetSize, height: UpNextListItemViewConstants.height)
+            .contentShape(Rectangle())
         }
         .buttonStyle(.plain)
         .disabled(isUpdating)
         .frame(maxHeight: .infinity)
-        .padding(.trailing, theme.spacing.medium)
+        .padding(.trailing, theme.spacing.small)
     }
 }
 
-private enum WatchListItemViewConstants {
+private enum UpNextListItemViewConstants {
     static let height: CGFloat = 140
     static let imageWidth: CGFloat = 120
     static let cornerRadius: CGFloat = 2
-    static let checkmarkSize: CGFloat = 32
+    static let checkmarkSize: CGFloat = 36
+    static let tapTargetSize: CGFloat = 48
 }
 
 #Preview {
     VStack {
-        WatchListItemView(
+        UpNextListItemView(
             episode: SwiftNextEpisode(
                 showTraktId: 1,
                 showName: "The Walking Dead: Daryl Dixon",
@@ -157,10 +175,10 @@ private enum WatchListItemViewConstants {
                 runtime: "45 min",
                 overview: "Daryl washes ashore in France.",
                 badge: .premiere,
-                remainingEpisodes: 7
+                remainingEpisodes: 7,
+                watchedCount: 3,
+                totalCount: 10
             ),
-            premiereLabel: "PREMIERE",
-            newLabel: "NEW",
             onItemClicked: { _, _ in },
             onShowTitleClicked: { _ in },
             onMarkWatched: {}
